+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/StackExchangeRedisInstrumentationOptions.cs
@@ -3,7 +3,6 @@
 
 using System.Diagnostics;
 using OpenTelemetry.Trace;
-using StackExchange.Redis.Profiling;
 
 namespace OpenTelemetry.Instrumentation.StackExchangeRedis;
 
@@ -23,26 +22,26 @@ public class StackExchangeRedisInstrumentationOptions
     public bool SetVerboseDatabaseStatements { get; set; }
 
     /// <summary>
-    /// Gets or sets a callback method allowing to filter out particular <see cref="IProfiledCommand"/> instances.
+    /// Gets or sets a callback method allowing to filter out particular <see cref="RedisInstrumentationContext"/> instances.
     /// </summary>
     /// <remarks>
-    /// The filter callback receives the <see cref="IProfiledCommand"/> for the
+    /// The filter callback receives the <see cref="RedisInstrumentationContext"/> for the
     /// processed profiled command and returns a boolean indicating whether it should be filtered out.
     /// <list type="bullet">
     /// <item>If filter returns <see langword="true"/> the event is collected.</item>
     /// <item>If filter returns <see langword="false"/> or throws an exception the event is filtered out (NOT collected).</item>
     /// </list>
     /// </remarks>
-    public Func<IProfiledCommand, bool>? Filter { get; set; }
+    public Func<RedisInstrumentationContext, bool>? Filter { get; set; }
 
     /// <summary>
     /// Gets or sets an action to enrich an Activity.
     /// </summary>
     /// <remarks>
     /// <para><see cref="Activity"/>: the activity being enriched.</para>
-    /// <para><see cref="IProfiledCommand"/>: the profiled redis command from which additional information can be extracted to enrich the activity.</para>
+    /// <para><see cref="RedisInstrumentationContext"/>: the profiled redis command from which additional information can be extracted to enrich the activity.</para>
     /// </remarks>
-    public Action<Activity, IProfiledCommand>? Enrich { get; set; }
+    public Action<Activity, RedisInstrumentationContext>? Enrich { get; set; }
 
     /// <summary>
     /// Gets or sets a value indicating whether the <see cref="StackExchangeRedisConnectionInstrumentation"/> should enrich Activity with <see cref="ActivityEvent"/> entries about the Redis command processing/lifetime. Defaults to <see langword="true"/>.